+++ src/main/java/dev/vesper/eveningstarlib/EveningStarLib.java
@@ -1,6 +1,5 @@
 package dev.vesper.eveningstarlib;
 
-import dev.vesper.eveningstarlib.common.Aurora.AuroraDefaultUniforms;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;