+++ src/main/java/dev/vesper/eveningstarlib/common/Aurora/Aurora.java
@@ -1,13 +1,18 @@
 package dev.vesper.eveningstarlib.common.Aurora;
 
 
+import com.google.common.annotations.Beta;
 import io.netty.util.internal.UnstableApi;
 
 import java.util.HashMap;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 
+// WIP disabled in production until its more ready :)
+
+/*
 @UnstableApi
+@Beta
 public class Aurora {
 
     private static final Map<String, UniformData> uniforms = new ConcurrentHashMap<>();
@@ -116,4 +121,4 @@ public class Aurora {
             return glslType;
         }
     }
-}
+}*/