+++ src/main/java/dev/vesper/eveningstarlib/common/ESLModChecks.java
@@ -2,8 +2,8 @@ package dev.vesper.eveningstarlib.common;
 
 //? fabric {
 import net.fabricmc.loader.api.FabricLoader;
-//?}
 import net.irisshaders.iris.api.v0.IrisApi;
+//?}
 //? neoforge {
 /*import net.neoforged.fml.ModList;
 *///?}
@@ -18,7 +18,7 @@ public class ESLModChecks {
         /*return ModList.get().isLoaded("iris");
         *///?}
     }
-
+//? fabric {
     public static boolean isShaders() {
         if (isIris()){
             return IrisApi.getInstance().getConfig().areShadersEnabled();
@@ -26,6 +26,7 @@ public class ESLModChecks {
             return false;
         }
     }
+    //?}
 
     public static boolean isSodium(){
         //? fabric {