+++ src/main/java/dev/vesper/eveningstarlib/fabric/FabricClientEntrypoint.java
@@ -2,7 +2,6 @@ package dev.vesper.eveningstarlib.fabric;
 
 //? fabric {
 import dev.vesper.eveningstarlib.EveningStarLib;
-import dev.vesper.eveningstarlib.common.Aurora.AuroraDefaultUniforms;
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.fabric.api.client.rendering.v1.level.LevelRenderEvents;
 
@@ -12,9 +11,9 @@ public class FabricClientEntrypoint implements ClientModInitializer {
     public void onInitializeClient() {
         EveningStarLib.LOG.info("Initializing {} Client", EveningStarLib.MOD_ID);
 
-        AuroraDefaultUniforms.registerAll();
-        // Not sure if this is the right event, was BEFORE_ENTITIES before the FAPI changes
-        LevelRenderEvents.AFTER_OPAQUE_TERRAIN.register(worldRenderContext -> {AuroraDefaultUniforms.updateAll();});
+        // AuroraDefaultUniforms.registerAll();
+        // Not sure if this is the right event, was BEFORE_ENTITIES before the FAPI changes. Just picked the one that felt most right needs testing
+        // LevelRenderEvents.AFTER_TRANSLUCENT_FEATURES.register(worldRenderContext -> {AuroraDefaultUniforms.updateAll();});
     }
 
 }