+++ src/main/java/dev/vesper/eveningstarlib/mixin/GameRenderMixin.java
@@ -1,7 +1,6 @@
 package dev.vesper.eveningstarlib.mixin;
 
 import com.mojang.blaze3d.systems.RenderSystem;
-import dev.vesper.eveningstarlib.common.Aurora.AuroraUploader;
 import net.minecraft.client.DeltaTracker;
 import net.minecraft.client.renderer.GameRenderer;
 import org.spongepowered.asm.mixin.Mixin;
@@ -15,16 +14,16 @@ public class GameRenderMixin {
     @Inject(method = "renderLevel", at = @At("HEAD"))
     public void renderLevel(DeltaTracker deltaTracker, CallbackInfo ci) {
         RenderSystem.assertOnRenderThread();
-        AuroraUploader.uploadUniformsToCurrentProgram();
+        //AuroraUploader.uploadUniformsToCurrentProgram();
     }
 
     @Inject(method = "resetData", at = @At("RETURN"))
     private void afterResetData(CallbackInfo ci){
-        AuroraUploader.clearCache();
+        //AuroraUploader.clearCache();
     }
 
     @Inject(method = "resize", at = @At("RETURN"))
     private void afterResize(CallbackInfo ci){
-        AuroraUploader.clearCache();
+        //AuroraUploader.clearCache();
     }
 }